Removed clearing of Cart
We simply don't understand HOW to link a button to a function that does a delete request to /rest/cart to then remove the cart with the current id from the database.

And yes, most of those steps have been used in this project in one way or another, but due to the lack of basic understanding of the different components 3 of us togehter couldn't figure out how to reapply the same steps to make them do what we needed them to.

I got as far as to having a button linked to a function that does a delete request, but you can't send a req.body through a delete (something I didn't know) so without having to create links, and showing the current cart id in the url, using req.params instead. We can't create a correct delete request.
Could've been possible to do it with a post, but that's not what post is made for.